First Pause Menu Implementation
Added to Forest 1-1 and added a Controls section to Main Menu the Pause Menu works MOSTLY there is one issue that I will address as soon as possible since prior I couldn't find something that might be releated to the problem. Minus that slight problem the Pause Menu is operated by using the escape button and again will talk more about it soon.